- rename the property handlers from set => setp, get => getp, - added remove property handler rmp - remove code which was looking up dlls in the gac to resolve namespace includes, turns out csc is smart enough to resolve System.Xml for example to the correct dlls. gg csc.exe
- change the way task are handle, task are no longer the action to be execute but
instead the data needed for executing a task, each task type has a handler that can
execute it, this makes generating reusable task so much easier.

common code can be capture and parametarized in a task object and then the task handler
will execute it. This frees the user from having to worry if code in a task fail or pass
and the user just provides what he wants done and the task runner will take care of
the rest.